move internal/codeintel/lsif/correlation => lib/codeintel/lsif/conversion
