+++ enterprise/lib/codeintel/lsif/conversion/canonicalize.go
@@ -1,4 +1,4 @@
-package correlation
+package conversion
 
 import (
 	"sort"
