+++ enterprise/lib/codeintel/lsif/conversion/canonicalize_test.go
@@ -1,4 +1,4 @@
-package correlation
+package conversion
 
 import (
 	"testing"
