+++ enterprise/lib/codeintel/lsif/conversion/errors.go
@@ -1,4 +1,4 @@
-package correlation
+package conversion
 
 import (
 	"errors"
