+++ enterprise/lib/codeintel/lsif/conversion/group.go
@@ -1,4 +1,4 @@
-package correlation
+package conversion
 
 import (
 	"context"
