+++ enterprise/lib/codeintel/lsif/conversion/prune.go
@@ -1,4 +1,4 @@
-package correlation
+package conversion
 
 import (
 	"context"
