+++ enterprise/lib/codeintel/lsif/conversion/query.go
@@ -1,4 +1,4 @@
-package correlation
+package conversion
 
 import (
 	"errors"
