+++ enterprise/lib/codeintel/lsif/conversion/util.go
@@ -1,4 +1,4 @@
-package correlation
+package conversion
 
 import (
 	"sort"
